refactor: 구간 삭제 기능 리팩토링
